fugue::subsonic

Module models

Constants

ID_FIELDS 🔒

The list of JSON keys that contain Subsonic IDs and should be namespaced.

Traits

NamespaceIds

Trait to rewrite all Subsonic IDs in a JSON value to be namespaced.

Functions

merge_album_lists

Merge album lists from multiple backends.

merge_artist_indexes

Extract and merge artist index arrays from multiple backend responses. Each backend returns {“artists”: {“index”: [{“name”: “A”, “artist”: […]}, …]}}. We merge indexes by letter and combine artist lists.

merge_playlists

Merge playlists from multiple backends.

merge_random_songs

Merge random songs from multiple backends.

merge_search_results

Merge search results from multiple backends.

merge_starred

Merge starred items from multiple backends.